Update to latest library PR revision
